+++ rocAL/rocAL_pybind/amd/rocal/plugin/tf.py
@@ -113,7 +113,7 @@ class ROCALGenericIteratorDetection(object):
             self.labels = np.zeros(self.count_batch, dtype="int32")
             self.loader.GetBBLabels(self.labels)
             # 1D bboxes array in a batch
-            self.bboxes = np.zeros((self.count_batch*4), dtype="float32")
+            self.bboxes = np.zeros((self.count_batch*4), dtype="float64")
             self.loader.GetBBCords(self.bboxes)
             #1D Image sizes array of image in a batch
             self.img_size = np.zeros((self.bs * 2),dtype = "int32")
